# colexec: inline datum conversion in EncDatumRowsToColVec