… pkg/sql/colexec/rowstovec_tmpl.go
… pkg/sql/colexec/select_in.eg.go
… pkg/sql/colexec/select_in_tmpl.go
… pkg/sql/colexec/selection_ops.eg.go
… pkg/sql/colexec/selection_ops_tmpl.go
… pkg/sql/colexec/utils.eg.go
… pkg/sql/colexec/utils.go
… pkg/sql/colexec/utils_tmpl.go