**colexec: remove error return parameter for GetDatumToPhysicalFn**

GetDatumToPhysicalFn returns a function for converting a datum of the given
ColumnType to the corresponding Go type. Now the signature of the return
function no longer contains an error since we assume that the conversion
must succeed. If for some reason it fails, a panic will be emitted and will
be caught by the panic-catcher mechanism of the vectorized engine and will
be propagated as an error accordingly.

The removal of the error return parameter seems to give some performance
improvement.

Release note: None

**colexec: inline datum conversion in EncDatumRowsToColVec**

Previously, `EncDatumRowsToColVec` function internally used another
converter function which doesn't get inlined. This commit adjusts the
template to remove one level of funciton "indirection" giving some
performance improvement.

Additionally, this commit reuses infrastructure of `rowsToVec` template
to generate GetDatumToPhysicalFn. This allows us to have the conversion
defined in a single place.

Release note: None