@yuzefovich 2020-07-03
Benchmark of the first commit against master:
```
name                       old time/op    new time/op    delta
Columnarize-16                475µs ± 1%     453µs ± 2%  -4.59%  (p=0.000 n=10+10)
ColumnarizeMaterialize-16     790µs ± 1%     757µs ± 1%  -4.13%  (p=0.000 n=10+9)

name                       old speed      new speed      delta
Columnarize-16              337MB/s ± 1%   353MB/s ± 2%  +4.81%  (p=0.000 n=10+10)
ColumnarizeMaterialize-16   203MB/s ± 1%   211MB/s ± 1%  +4.31%  (p=0.000 n=10+9)

name                       old alloc/op   new alloc/op   delta
Columnarize-16                161kB ± 0%     161kB ± 0%  -0.02%  (p=0.000 n=10+9)
ColumnarizeMaterialize-16     357kB ± 0%     357kB ± 0%  -0.01%  (p=0.000 n=10+10)

name                       old allocs/op  new allocs/op  delta
Columnarize-16                20.0k ± 0%     20.0k ± 0%  -0.00%  (p=0.000 n=9+8)
ColumnarizeMaterialize-16     21.3k ± 0%     21.3k ± 0%  -0.00%  (p=0.046 n=8+10)
```

Benchmark of both commits against master:
```
name                       old time/op    new time/op    delta
Columnarize-16                476µs ± 1%     430µs ± 2%   -9.70%  (p=0.000 n=10+10)
ColumnarizeMaterialize-16     790µs ± 1%     722µs ± 0%   -8.61%  (p=0.000 n=10+10)

name                       old speed      new speed      delta
Columnarize-16              336MB/s ± 1%   372MB/s ± 2%  +10.75%  (p=0.000 n=10+10)
ColumnarizeMaterialize-16   202MB/s ± 1%   222MB/s ± 0%   +9.42%  (p=0.000 n=10+10)

name                       old alloc/op   new alloc/op   delta
Columnarize-16                161kB ± 0%     161kB ± 0%   -0.04%  (p=0.000 n=10+10)
ColumnarizeMaterialize-16     357kB ± 0%     357kB ± 0%   -0.03%  (p=0.000 n=10+10)

name                       old allocs/op  new allocs/op  delta
Columnarize-16                20.0k ± 0%     20.0k ± 0%   -0.00%  (p=0.000 n=9+9)
ColumnarizeMaterialize-16     21.3k ± 0%     21.3k ± 0%   -0.01%  (p=0.000 n=10+10)
```